move sum& avg tests, add new indexes required
